+++ src/components/FormComponents/ConfirmPassword/ConfirmPasswordForm.tsx
@@ -10,7 +10,8 @@ import { toastFail, toastSuccess } from "@nepMeds/components/Toast";
 import { NAVIGATION_ROUTES } from "@nepMeds/routes/routes.constant";
 import { useResetPasswordMutation } from "@nepMeds/service/nepmeds-forgot-password";
 import { colors } from "@nepMeds/theme/colors";
-import { Navigate, useParams } from "react-router-dom";
+import { Navigate, useNavigate, useParams } from "react-router-dom";
+import { AxiosError } from "axios";
 
 const schema = yup.object().shape({
   password: yup.string().required("Password is required"),
@@ -18,8 +19,12 @@ const schema = yup.object().shape({
 });
 
 const ConformPasswordForm = () => {
-  const { uid = "", token = "" } = useParams<{ uid: string; token: string }>();
+  const { uidb64 = "", token = "" } = useParams<{
+    uidb64: string;
+    token: string;
+  }>();
   const resetPasswordAction = useResetPasswordMutation();
+  const navigate = useNavigate();
 
   const [showPassword, setShowPassword] = useState(false);
   const [showConfirmPassword, setShowConfirmPassword] = useState(false);
@@ -47,19 +52,29 @@ const ConformPasswordForm = () => {
 
   const onSubmit = async () => {
     try {
-      await resetPasswordAction.mutateAsync({
-        token,
-        uid,
-        new_password: getValues("password"),
-        confirm_new_password: getValues("confirmPassword"),
-      });
-      toastSuccess("New password saved successfully!");
+      await resetPasswordAction
+        .mutateAsync({
+          token,
+          uidb64,
+          new_password: getValues("password"),
+          confirm_new_password: getValues("confirmPassword"),
+        })
+        .then(() => {
+          toastSuccess("New password saved successfully!");
+          navigate("/login");
+        })
+        .catch(error => {
+          {
+            const err = error as AxiosError<{ message: string }>;
+            toastFail(err?.response?.data?.message || "");
+          }
+        });
     } catch (error) {
       toastFail("Failed to update password!");
     }
   };
 
-  if (!uid && !token) {
+  if (!uidb64 && !token) {
     return <Navigate to={NAVIGATION_ROUTES.LOGIN} replace />;
   }